+++ gulpfile.js
@@ -2,6 +2,7 @@ var gulp = require('gulp'),
 
 	// general
 	watch = require('gulp-watch'),
+	runSequence = require('run-sequence'),
 	livereload = require('gulp-livereload'),
 	rename = require('gulp-rename')
 	concat = require('gulp-concat'),
@@ -164,7 +165,18 @@ gulp.task('pattern-scripts', function () {
 // TASKS
 
 // post deploy, for deploying to the server
-gulp.task('post-deploy', ['templates', 'scss', 'js', 'modernizr']);
+gulp.task('post-deploy', function(callback) {
+
+	// use run sequence to make sure modernizr isn't called prematurely
+	runSequence(
+		['templates', 'scss', 'js'],
+		'modernizr',
+		callback
+	);
+
+});
+
+gulp.task('default', ['watch']);
 
 // default watch, for normal development
 gulp.task('default', ['watch']);